move test_dirs fixture to conftest.py
